Fix multichain swapout method
