+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "rubic-sdk",
-  "version": "4.0.0-beta.3",
+  "version": "4.0.1-alpha-multichain.1",
   "description": "Simplify dApp creation",
   "main": "lib/index.js",
   "types": "lib/index.d.ts",
